+++ app/models/user.rb
@@ -1,3 +1,5 @@
+require "securerandom"
+
 class User < ApplicationRecord
   has_many :lists, dependent: :destroy
   has_many :kanban_boards, dependent: :destroy
@@ -13,6 +15,7 @@ class User < ApplicationRecord
   has_many :agent_runs, dependent: :destroy
   has_many :agent_events, dependent: :destroy
   has_many :agent_recommendations, dependent: :destroy
+  has_many :activity_logs, dependent: :destroy
 
   # Single-user application - one user per instance
   validates :email, presence: true, uniqueness: true
@@ -93,6 +96,9 @@ class User < ApplicationRecord
   DEFAULT_VOICE_ID_SETTINGS = {
     'enabled' => false
   }.freeze
+  DEFAULT_MOBILE_UPLINK_SETTINGS = {
+    'enabled' => false
+  }.freeze
   DEFAULT_IDEA_WORK_TOKEN_SETTINGS = {
     'enabled' => false
   }.freeze
@@ -110,6 +116,7 @@ class User < ApplicationRecord
   MAX_TYPING_LOCK_SECONDS = 24.hours.to_i
   MIN_TYPING_LOCK_FAILED_UNLOCK_COOLDOWN_SECONDS = 1.minute.to_i
   MAX_TYPING_LOCK_FAILED_UNLOCK_COOLDOWN_SECONDS = 24.hours.to_i
+  MOBILE_UPLINK_ID_BYTES = 18
 
   ALLOWED_NOTIFICATION_TRIGGERS = %w[
     state_changed score_changed added_to_list created
@@ -323,6 +330,7 @@ class User < ApplicationRecord
   def update_scoring_weights(weights)
     self.settings ||= {}
     self.settings['scoring_weights'] = weights.slice('trl', 'difficulty', 'opportunity', 'timing')
+                                              .transform_values(&:to_f)
     save!
   end
 
@@ -567,6 +575,43 @@ class User < ApplicationRecord
     self.settings['voice_id'] ||= {}
     self.settings['voice_id'].delete('fingerprint')
     self.settings['voice_id'].delete('fingerprint_ciphertext')
+    save
+  end
+
+  def mobile_uplink_settings
+    DEFAULT_MOBILE_UPLINK_SETTINGS.merge(settings&.dig('mobile_uplink') || {})
+  end
+
+  def mobile_uplink_enabled?
+    ActiveModel::Type::Boolean.new.cast(mobile_uplink_settings['enabled']) == true && mobile_uplink_configured?
+  end
+
+  def mobile_uplink_configured?
+    mobile_uplink_id.present?
+  end
+
+  def mobile_uplink_id
+    mobile_uplink_settings['connect_id'].presence
+  end
+
+  def mobile_uplink_pairing_payload(workspace_url:)
+    return unless mobile_uplink_id
+
+    MobileUplink.pairing_payload(uplink_id: mobile_uplink_id, workspace_url: workspace_url)
+  end
+
+  def update_mobile_uplink_settings(params)
+    enabled = ActiveModel::Type::Boolean.new.cast(params.fetch('enabled', false)) == true
+    self.settings ||= {}
+    self.settings['mobile_uplink'] ||= {}
+
+    if enabled
+      self.settings['mobile_uplink']['enabled'] = true
+      self.settings['mobile_uplink']['connect_id'] ||= SecureRandom.urlsafe_base64(MOBILE_UPLINK_ID_BYTES)
+    else
+      self.settings['mobile_uplink'] = { 'enabled' => false }
+    end
+
     save
   end